+++ CHANGELOG.md
@@ -96,6 +96,7 @@
 - Fixed a redirect bug with oauth
 - Removed support for Mozilla Persona
 - elastic.scroll does not return nil sid when there are no more results (#315)
+- stats.lua should return firstMonth and lastMonth (#120)
 
 ## CHANGES in 0.9b: